requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
pypdf2==3.0.1
python-docx==0.8.11
markdown==3.5.1